requirement.txt
arrow==0.15.8
asgiref==3.2.10
blessed==1.17.8
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
croniter==0.3.34
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.0
django-elasticsearch-dsl==7.1.1
django-extensions==3.0.1
django-phone-field==1.8.1
django-picklefield==3.0.1
django-q==1.3.2
django-rest-knox==4.1.0
django-sslserver==0.22
djangorestframework==3.11.1
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
idna==2.9
mysqlclient==1.4.6
natsort==7.0.1
oauthlib==3.1.0
Pillow==7.1.2
psutil==5.7.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
